+++ src/cli.ts
@@ -1,13 +1,12 @@
 #!/usr/bin/env node
 
 import minimist from 'minimist'
-import {readFile, writeFile, existsSync, lstatSync, readdirSync} from 'mz/fs'
-import * as mkdirp from 'mkdirp'
+import {readFileSync, writeFileSync, existsSync, lstatSync, readdirSync, mkdirSync} from 'fs'
 import {glob} from 'glob'
 import isGlob from 'is-glob'
-import {join, resolve, dirname, basename} from 'path'
+import {join, resolve, dirname} from 'path'
 import {compile, DEFAULT_OPTIONS, Options} from './index'
-import {pathTransform, error} from './utils'
+import {pathTransform, error, parseFileAsJSONSchema, justName} from './utils'
 
 main(
   minimist(process.argv.slice(2), {
@@ -88,7 +87,7 @@ async function processGlob(argIn: string, argOut: string | undefined, argv: Part
 
   // careful to do this serially
   results.forEach(([file, result]) => {
-    const outputPath = argOut && `${argOut}/${basename(file, '.json')}.d.ts`
+    const outputPath = argOut && `${argOut}/${justName(file)}.d.ts`
     outputResult(result, outputPath)
   })
 }
@@ -110,23 +109,24 @@ async function processDir(argIn: string, argOut: string | undefined, argv: Parti
 
   // careful to do this serially
   results.forEach(([file, result, outputPath]) =>
-    outputResult(result, outputPath ? `${outputPath}/${basename(file, '.json')}.d.ts` : undefined),
+    outputResult(result, outputPath ? `${outputPath}/${justName(file)}.d.ts` : undefined),
   )
 }
 
-async function outputResult(result: string, outputPath: string | undefined): Promise<void> {
+function outputResult(result: string, outputPath: string | undefined): void {
   if (!outputPath) {
     process.stdout.write(result)
   } else {
     if (!isDir(dirname(outputPath))) {
-      mkdirp.sync(dirname(outputPath))
+      mkdirSync(dirname(outputPath), {recursive: true})
     }
-    return await writeFile(outputPath, result)
+    return writeFileSync(outputPath, result)
   }
 }
 
 async function processFile(argIn: string, argv: Partial<Options>): Promise<string> {
-  const schema = JSON.parse(await readInput(argIn))
+  const {filename, contents} = await readInput(argIn)
+  const schema = parseFileAsJSONSchema(filename, contents)
   return compile(schema, argIn, argv)
 }
 
@@ -140,15 +140,21 @@ function getPaths(path: string, paths: string[] = []) {
   return paths
 }
 
-async function readInput(argIn?: string): Promise<string> {
+async function readInput(argIn?: string): Promise<{filename: string | null; contents: string}> {
   if (!argIn) {
-    return readStream(process.stdin)
+    return {
+      filename: null,
+      contents: await readStream(process.stdin),
+    }
   }
-  return readFile(resolve(process.cwd(), argIn), 'utf-8')
+  return {
+    filename: argIn,
+    contents: readFileSync(resolve(process.cwd(), argIn), 'utf-8'),
+  }
 }
 
 async function readStream(stream: NodeJS.ReadStream): Promise<string> {
-  const chunks = []
+  const chunks: Uint8Array[] = []
   for await (const chunk of stream) chunks.push(chunk)
   return Buffer.concat(chunks).toString('utf8')
 }
@@ -176,6 +182,8 @@ Boolean values can be set to false using the 'no-' prefix.
       Declare external schemas referenced via '$ref'?
   --enableConstEnums
       Prepend enums with 'const'?
+  --inferStringEnumKeysFromValues
+      Create enums from JSON enums instead of union types
   --format
       Format code? Set this to false to improve performance.
   --maxItems